Fix several unit tests that seemed not to test what was intended
There were several (= x y) expressions changed to (is (= x y)).  They
had no is, and thus could never cause tests to fail.

Changed several occurrences of (is expr1 expr2) to (is (= expr1
expr2)).  When expr1 evaluates to a value that is neither false nor
nil, the former kind of 'test' always passes.

Similarly for occurrences of (is (= expr1) expr2).  (= expr1) is
always true, so the test always passes.

Found using a pre-release version of the Eastwood Clojure lint tool.

I apologize for not knowing enough about ogre to fix the tests that
now do not pass.  I was hoping an ogre developer could look into
those, now that they can see which ones are failing.